+++ spec/atrea_control/duplex_spec.rb
@@ -38,6 +38,7 @@ RSpec.describe AtreaControl::Duplex do
     before do
       duplex.instance_variable_set :@logged, true
       duplex.driver.get "file://#{File.join(__dir__, "../fixtures/files/dashboard.html")}"
+      duplex.send(:refresh!)
     end
 
     describe "#name" do